requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.4.0
pydantic[email]>=2.4.0
playwright>=1.40.0
python-dotenv>=1.0.0

# LLM Providers
google-generativeai>=0.3.0
openai>=1.3.0
anthropic>=0.7.0

# HTTP and Web
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=4.9.0

# Data Processing
pandas>=2.1.0
openpyxl>=3.1.0
pdfplumber>=0.10.0
PyPDF2>=3.0.0

# Visualization
matplotlib>=3.8.0
plotly>=5.18.0
kaleido>=0.2.1
pillow>=10.1.0

# Utilities
python-multipart>=0.0.6
